+++ examples/tcp_echo_server.py
@@ -9,22 +9,23 @@
     :license: BSD, see LICENSE for more details.
 """
 import time
+from typing import Optional
 
-from proxy.core.acceptor import AcceptorPool
 from proxy.proxy import Proxy
+from proxy.core.acceptor import AcceptorPool
+from proxy.core.base import BaseTcpServerHandler
 
-from examples.base_server import BaseServerHandler
 
-
-class EchoServerHandler(BaseServerHandler):  # type: ignore
+class EchoServerHandler(BaseTcpServerHandler):
     """Sets client socket to non-blocking during initialization."""
 
     def initialize(self) -> None:
         self.client.connection.setblocking(False)
 
-    def handle_data(self, data: memoryview) -> None:
+    def handle_data(self, data: memoryview) -> Optional[bool]:
         # echo back to client
         self.client.queue(data)
+        return None
 
 
 def main() -> None:
